+++ libs/eslint-config/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Change the rule import/extensions to support imorting files without extensions ([#92](https://github.com/vickev/howdypix/pull/92))
+
 ## [0.0.2] - 2020-03-07
 
 ### Added